:art: add ListFilterBar component
extracts the FileList Filters to their own component
