+++ src/App.css
@@ -24,6 +24,12 @@ body {
   margin: 0px !important;
 }
 
+.noVerticalPadding {
+  padding-top: 0px !important;
+  padding-bottom: 0px !important;
+
+}
+
 .noHorizontalPadding {
   padding-left: 0px !important;
   padding-right: 0px !important;
